--- client/pages/admin/font.tsx
@@ -1,6 +1,0 @@
-import Admin, { Type } from 'features/Admin';
-import React from "react";
-
-export default function AdminFontPage() {
-  return <Admin type={Type.Font} />
-}
