+++ client/pages/admin/font/index.tsx
@@ -1,0 +1,7 @@
+import { FontForm, Overview } from 'features/Admin/Font';
+import { useRouter } from 'next/router';
+import React from 'react';
+
+export default function Font() {
+  return <Overview />;
+}
