+++ client/src/features/Admin/Font/Form/index.ts
@@ -1,0 +1,3 @@
+import FontForm from './FontForm';
+
+export default FontForm;
